refactor(test): separate Fare tests for each decorator
